+++ Front-End/package.json
@@ -17,7 +17,7 @@
     "build": "npm run postinstall && npm run electron:tsc && ng build",
     "build:dev": "npm run build -- -c dev",
     "build:prod": "npm run build -- -c production",
-    "ng:serve": "ng serve --disable-host-check",
+    "ng:serve": "ng serve --source-map=false",
     "electron:tsc": "tsc",
     "electron:serve": "npm run electron:tsc && electron . --serve",
     "electron:mac": "electron-packager . --overwrite --platform=darwin --arch=x64 --prune=true --out=release-builds --icon=\"favicon.ico\"",
@@ -79,6 +79,7 @@
     "karma-jasmine": "~4.0.0",
     "karma-jasmine-html-reporter": "^1.5.0",
     "karma-phantomjs-launcher": "^1.0.4",
+    "lite-server": "^2.6.1",
     "mocha": "^8.2.1",
     "npm-run-all": "^4.1.3",
     "puppeteer": "^2.1.1",
