+++ Front-End/src/app/pages/home/home.component.scss
