+++ Front-End/src/app/pages/mode/mode.component.html
