+++ Front-End/src/app/pages/settings/settings.component.scss
