+++ Front-End/src/app/pages/visualizer/visualizer.component.scss
