+++ Front-End/src/app/shared/components/navigationbar/navigationbar.component.scss
@@ -1,4 +1,4 @@
-@import '../../../variables';
+@import '../../../../variables';
 
 //noinspection CssUnknownProperty
 #menu-bar {
